layers are floats now, better camera implementation!

+++ src/main/java/xyz/valnet/engine/scenegraph/IMouseCaptureArea.java
@@ -9,5 +9,5 @@ public interface IMouseCaptureArea {
   public void mouseUp(int button);
 
   public Vector4f getBox();
-  public int getLayer();
+  public float getLayer();
 }
